feat: Added data from json using The Movie Db website.
Now the website displays the list of most popular movies using the json data file. However this is static data and will need to modify the javascript code to add trailers and updated data every time the website is opened via. python file or using html. Also, there is an issue with UI; some rows of movie are displaying only 2 movies instead of 3.

Up next add code for trailer to work and data to get updated.